Develop

+++ client/src/pages/RegisterPage.jsx
@@ -1,29 +1,38 @@
-import { useState } from "react";
+import { useContext, useState } from "react";
 import { registerUser } from "../api/todolist.api";
 import { useNavigate } from "react-router-dom";
+import { AuthContext } from "../context/AuthContext";
 
 export function RegisterPage() {
-
+  const { login } = useContext(AuthContext);
   const navigate = useNavigate();
 
   const [username, setUsername] = useState("");
   const [password, setPassword] = useState("");
+  const [confirmPass, setConfirmPass] = useState("");
   const [email, setEmail] = useState("");
   const [error, setError] = useState("");
 
   const handleSingIn = async () => {
+
+    if (password != confirmPass) {
+      console.log("password didn't match");
+      return
+    }
+
     try {
 
       // console.log(username + " - " + password + " - " + email);
 
       const response = await registerUser(username, password, email);
-      
+      console.log(response);
       // console.log(response);
       if (response.status === 200) {
         const token = response.data.token;
         localStorage.setItem('token', token);  // Guardar el token en localStoreage
-        // console.log('Registration successful');
-        navigate('/home');  // Redirigir después del registro
+        console.log('Registration successful');
+        // navigate('/home');  // Redirigir después del registro
+        login(username, password);
       }
     } catch (error) {
       console.error("Signin error:", error.response);
@@ -47,6 +56,12 @@ export function RegisterPage() {
           placeholder="Password"
           value={password}
           onChange={(e) => setPassword(e.target.value)}
+        />
+        <input
+        type="password"
+        placeholder="Password"
+        value={confirmPass}
+        onChange={(e) => setConfirmPass(e.target.value)}
         />
         <input
           type="text"
